support custom processor for flatpage and listpage
