+++ components/FilesView.tsx
@@ -13,8 +13,7 @@ import AssetsManager from "../mex/assets/AssetsManager";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import stylesManager from "../mex/StylesManager";
 import { translate } from "../mex/assets/LocalizationManager";
-import {AttachmentMetadata} from "../Tools/MexModuleCodeGenerator/mex-engine-core/src/di/ServicesProxy";
-import {IconTypes} from "@skedulo/mex-engine-proxy";
+import {AttachmentMetadata, IconTypes} from "@skedulo/mex-engine-proxy";
 
 const FilesView = ({attachmentsMetadata, isSignature, readonly}: {
     attachmentsMetadata: AttachmentMetadata[],
