+++ mex/common/processors/page/ListPageProcessor.tsx
@@ -16,10 +16,8 @@ import MexAsyncText from "../../../../components/MexAsyncText";
 import StylesManager from "../../../StylesManager";
 import ThemeManager from "../../../colors/ThemeManager";
 import InternalUtils, {FilterByExpressionInfo, translateOneLevelOfExpression} from "../../InternalUtils";
-import {useCallback, useState, useRef, useMemo, useEffect, useReducer} from "react";
+import {useCallback, useState, useRef, useMemo} from "react";
 import {RightNavigationItem} from "../../../../components/NavigationItemsView";
-import * as lodash from "lodash";
-import { Dictionary } from 'lodash';
 import {ListPageSectionHeaderComponent} from "../../../../components/ListPage/ListPageSectionHeader";
 import {ListPageFooterComponent} from "../../../../components/ListPage/ListPageFooter";
 import {ListPageHeaderComponent} from "../../../../components/ListPage/ListPageHeader";
